+++ src/Group.js
@@ -1,8 +1,7 @@
 /**
  * @class Kartographer.Data.Group
  * @property {Object|Object[]|null} geoJSON
- * @property {string|null} name Empty for ExternalData, otherwise this is the
- *   group key in mapdata.
+ * @property {string|null} name The group key in mapdata.
  * @property {boolean} failed Flag is true if the group failed to fully load.
  * @property {Error|null} failureReason Details about a failure, if any.
  *